// client/src/lib/open-api/model/user-update.ts
/**
 * Diploma FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RoleEnum } from './role-enum';


export interface UserUpdate { 
    first_name: string;
    last_name: string;
    email: string;
    role: RoleEnum;
}
export namespace UserUpdate {
}